+++ src/pages/Content/index.jsx
@@ -1,11 +1,40 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import Frame from './modules/frame/frame';
+import UpdateNotice from './modules/UpdateNotice/UpdateNotice';
 
 let shouldShrinkBody = true;
 let sidebarLocation = 'left';
 const toggleButtonLocation = 'bottom';
 let sidebarWidth = 280;
+
+let updateNoticeRoot = document.createElement('div');
+document.body.appendChild(updateNoticeRoot);
+updateNoticeRoot.setAttribute('id', 'vt-update-notice');
+localStorage.removeItem('vt-update-notice-dismissed');
+
+document.addEventListener('visibilitychange', (info) => {
+  // console.log(document.hidden);
+  try {
+    chrome.runtime.sendMessage({ msg: 'GET_VERSION_NUMBER' }, (response) => {
+      // console.log(response);
+    });
+  } catch (err) {
+    if (err.toString().indexOf('Extension context invalidated') !== -1) {
+      if (!document.hidden) {
+        console.log('should refresh page');
+        ReactDOM.render(<UpdateNotice />, updateNoticeRoot);
+      } else {
+        ReactDOM.unmountComponentAtNode(updateNoticeRoot);
+      }
+    }
+  }
+});
+
+// let updateNoticeRoot = document.createElement('div');
+// document.body.appendChild(updateNoticeRoot);
+// updateNoticeRoot.setAttribute('id', 'vt-update-notice');
+// ReactDOM.render(<UpdateNotice />, updateNoticeRoot);
 
 const setSidebarWidth = (width) => {
   sidebarWidth = width;