+++ CHANGELOG.md
@@ -12,6 +12,12 @@ All notable changes to this project will be documented in this file.
 
 # v1.x (official releases)
 
+### [1.2.0](https://github.com/lxieyang/vertical-tabs-chrome-extension/releases/tag/v1.2.0) (2019-09-01)
+
+- Prompt users to refresh pages after the extension got updated to a newer version.
+
+![update-notice](preview/repo/update-notice.png)
+
 ### [1.1.0](https://github.com/lxieyang/vertical-tabs-chrome-extension/releases/tag/v1.1.0) (2019-09-01)
 
 - Support pinning/unpinning tabs in the context menu